Refactor mapper test packages to unify namespace
